pyproject.toml
[tool.poetry]
name = "wavelet-kit"
version = "0.1.0"
description = ""
authors = ["joypauls <joypaulsen3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
ipykernel = "^6.29.5"
matplotlib = "^3.9.2"
pillow = "^10.4.0"
numpy = "^2.1.1"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
